+++ packages/platform-browser/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @effect/platform-browser
 
+## 0.29.3
+
+### Patch Changes
+
+- Updated dependencies [[`6654f5f`](https://github.com/Effect-TS/effect/commit/6654f5f0f6b9d97165ede5e04ca16776e2599328), [`2eb11b4`](https://github.com/Effect-TS/effect/commit/2eb11b47752cedf233ef4c4395d9c4efc9b9e180), [`56c09bd`](https://github.com/Effect-TS/effect/commit/56c09bd369279a6a7785209d172739935818cba6), [`71aa5b1`](https://github.com/Effect-TS/effect/commit/71aa5b1c180dcb8b53aefe232d12a97bd06b5447), [`1700af8`](https://github.com/Effect-TS/effect/commit/1700af8af1131602887da721914c8562b6342393)]:
+  - effect@2.3.2
+  - @effect/platform@0.44.3
+
+## 0.29.2
+
+### Patch Changes
+
+- Updated dependencies [[`29739dd`](https://github.com/Effect-TS/effect/commit/29739dde8e6232824d49c4c7f8856de245249c5c)]:
+  - @effect/platform@0.44.2
+
 ## 0.29.1
 
 ### Patch Changes